Move accept into Tcp type
